Updated HUD
Updated HUD. Top bar displays resources, soldiers, and builder. Bottom has buttons for gameplay. The "Build Wall" button is currently set to do the same thing as "Build Tower" as I couldn't get either of those working due to an error, so that will need to be hooked up. I made the mini map slightly smaller so it would block less of the screen, however the border sometimes doesn't resize properly with the prefab.